fix & docs: Документирование и распределение на модули
