+++ build.gradle.kts
@@ -65,7 +65,7 @@ allProjects {
 }
 
 
-kotlin{
+kotlin {
     jvm("desktop")
     sourceSets {
         val commonMain by getting {
