+++ demos/desktop_test/src/main/kotlin/Main.kt
@@ -1,24 +1,7 @@
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.size
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.application
-import core.dsl.config
-
-//@Composable
-//@Preview
-//fun App() {
-//    var text by remember { mutableStateOf("Hello, World!") }
-//
-//    MaterialTheme {
-//        Button(onClick = {
-//            text = "Hello, Desktop!"
-//        }) {
-//            Text(text)
-//        }
-//    }
-//}
+import core.dsl.elements.configuration.config
 
 fun main() {
     application {
@@ -26,7 +9,6 @@ fun main() {
             icon {
                 Image(painterResource("Good Tick.svg"), null, modifier = it)
             }
-            this.name = "test123"
         }.render(::exitApplication)
     }
