+++ gradle.properties
@@ -17,5 +17,4 @@ android.minSdk=24
 #Versions
 kotlin.version=1.9.0
 compose.version=1.5.0
-
 multimark-version=0.0.1
