+++ src/commonMain/kotlin/App.kt
@@ -1,12 +1,8 @@
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import ui.screen.StartScreen
 
 @Composable
 fun App() {
-
-
     StartScreen()
-
 }
