+++ src/commonMain/kotlin/core/dsl/elements/Icon.kt
@@ -1,7 +1,6 @@
 package core.dsl.elements
 
 import androidx.compose.foundation.Image
-import androidx.compose.material.Icon
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.painterResource
 import core.dsl.ConfigurationTagMaker
@@ -10,6 +9,7 @@ import core.dsl.elements.template.Element
 @ConfigurationTagMaker
 class Icon : Element {
     var icon: @Composable () -> Unit = { Image(painterResource("logo.svg"), null) }
+
     @Composable
     override fun render() {
         icon()
