+++ src/commonMain/kotlin/core/dsl/elements/configuration/ConfigurationPlatform.kt
@@ -1,0 +1,15 @@
+package core.dsl.elements.configuration
+
+import androidx.compose.runtime.Composable
+
+/**
+ * Реализация нативных компонентов
+ *
+ * [render(onCloseRequest)] - отображение с учитыванием закрытия окна
+ * @author Панков Вася (pank-su)
+ */
+expect abstract class ConfigurationPlatform() : Configuration {
+
+    @Composable
+    open fun render(onCloseRequest: () -> Unit)
+}
